AccommodationController, DATA.SQL merge
